app: set NetChangeEvent to double

+++ src/TTX.App/Jobs/CreatorValues/NetChangeEvent.cs
@@ -2,4 +2,4 @@ using TTX.Domain.ValueObjects;
 
 namespace TTX.App.Jobs.CreatorValues;
 
-public record NetChangeEvent(ModelId CreatorId, int NetChange);
+public record NetChangeEvent(ModelId CreatorId, double NetChange);
